split types part 3
https://coveord.atlassian.net/browse/KIT-2600
